fixed amount in entities

+++ src/main/java/dev/university/degree/entities/AppointmentProcedures.java
@@ -19,5 +19,5 @@ public class AppointmentProcedures {
     private Procedure procedure;
     @ManyToOne
     private MedicationStorage medicationStorage;
-    private int medicationAmount;
+    private double medicationAmount;
 }
